Skipping results which are younger than 3 days

+++ src/main/java/booru/BoardType.java
@@ -107,7 +107,7 @@ public enum BoardType {
         return maxLimit;
     }
 
-    public String getPageUrl(int id) {
+    public String getPageUrl(long id) {
         return pagePrefix + id;
     }
